feat(directory): refactor directory part 3
