+++ src/app/views/movie-detail/interface/movie-props.interface.ts
@@ -1,3 +1,4 @@
+import { IDefaultPageProps } from '@/library/interface/general/base.interface';
 import { IMovieDetail } from '@/library/model/movie/interface/movie-detail.interface';
 
 /**
@@ -8,3 +9,10 @@ import { IMovieDetail } from '@/library/model/movie/interface/movie-detail.inter
 export interface IMovieDetailProps {
   movie?: IMovieDetail;
 }
+
+/**
+ * Default Movie Detail Props Interface
+ * @author Irfan Andriansyah <irfan@99.co>
+ * @since 2021.02.05
+ */
+export type IDefaultMovieDetailProps = IDefaultPageProps & IMovieDetailProps;
