+++ src/app/views/movie-detail/seo/index.ts
@@ -2,7 +2,7 @@
  * Generate Class
  * @author Irfan Andriansyah <irfan@99.co>
  * @description
- * @since 2021.02.03
+ * @since 2021.02.05
  */
 class NameClass {}
