+++ src/app/views/movie-detail/seo/meta-tag.ts
@@ -1,0 +1,9 @@
+/**
+ * Generate Class
+ * @author Irfan Andriansyah <irfan@99.co>
+ * @description
+ * @since 2021.02.05
+ */
+class NameClass {}
+
+export default NameClass;
