+++ src/app/views/movie-detail/server-side.ts
@@ -3,7 +3,8 @@ import { GetServerSideProps } from 'next';
 import { IMovieDetailProps } from '@/app/views/movie-detail/interface';
 import { VerifiedIsNotEmpty } from '@/library/helper/validator.helper';
 import { IMovieDetail as Detail } from '@/library/model/movie/interface/movie-detail.interface';
-import { csrfToken, setupWeb } from '@/library/modules/csrf';
+// import { csrfToken, setupWeb } from '@/library/modules/csrf';
+import { csrfToken } from '@/library/modules/csrf';
 import { Get } from '@/library/modules/fetch-api';
 import { IFetchAPIWithoutData } from '@/library/modules/fetch-api/interface/fetch-api.interface';
 import { INextAPIResponse as API } from '@/library/modules/next-api/interface/next-api.interface';
@@ -13,17 +14,19 @@ import { INextAPIResponse as API } from '@/library/modules/next-api/interface/ne
  * @param {GetServerSidePropsContext} context - context page
  * @returns {Promise<GetServerSidePropsResult>}
  */
-export const getServerSideProps: GetServerSideProps = async (param) => {
+export const getServerSideProps: GetServerSideProps = async ({
+  params: { movieID }
+}) => {
   /**
    * Get Movie Detail Props
    * @returns {Promise<IMovieDetailProps>}
    */
-  const GetMovieDetailProps = (movieID: string): Promise<IMovieDetailProps> =>
+  const GetMovieDetailProps = (id: string): Promise<IMovieDetailProps> =>
     (Get as IFetchAPIWithoutData<API<Detail>>)(
       `http://localhost:3000/api/movie/detail`,
       {
         params: {
-          id: movieID
+          id
         }
       }
     )
@@ -38,7 +41,11 @@ export const getServerSideProps: GetServerSideProps = async (param) => {
       })
       .catch(() => ({}));
 
-  return setupWeb(async ({ params: { movieID } }) => ({
+  // return setupWeb(async ({ params: { movieID } }) => ({
+  //   props: { ...(await GetMovieDetailProps(`${movieID}` || ``)), csrfToken }
+  // }))(param);
+
+  return {
     props: { ...(await GetMovieDetailProps(`${movieID}` || ``)), csrfToken }
-  }))(param);
+  };
 };
