+++ src/library/interface/general/seo.interface.ts
@@ -1,0 +1,10 @@
+/**
+ * SEO Props Interface
+ * @author Irfan Andriansyah <irfan@99.co>
+ * @since 2021.02.05
+ */
+export interface ISeoProps {
+  linkTag?: string[];
+  metaTag?: string[];
+  title?: string;
+}
