+++ src/library/interface/language/language.interface.ts
@@ -1,0 +1,6 @@
+/**
+ * Language Preset
+ * @author Irfan Andriansyah <irfan@99.co>
+ * @since 2021.02.05
+ */
+export type ILanguagePreset = 'id' | 'en';
