+++ src/pages/movie/detail/[movieID].tsx
@@ -1,2 +1,2 @@
-export { getServerSideProps } from '@/app/views/movie-detail/app';
-export { default } from '@/app/views/movie-detail/container';
+export { getServerSideProps } from '@/app/views/movie-detail/server-side';
+export { default } from '@/app/views/movie-detail/client-side';
